+++ src/fastoad/gui/tests/test_analysis_and_plots.py
@@ -15,6 +15,7 @@ Tests for analysis and plots functions
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from pathlib import Path
+import pytest
 
 from .. import (
     aircraft_geometry_plot,
@@ -28,12 +29,14 @@ from .. import (
 DATA_FOLDER_PATH = Path(__file__).parent / "data"
 
 
-def test_wing_geometry_plot():
+@pytest.mark.parametrize(
+    "filename",
+    [DATA_FOLDER_PATH / "problem_outputs.xml", DATA_FOLDER_PATH / "problem_outputs_updated.xml"],
+)
+def test_wing_geometry_plot(filename):
     """
     Basic tests for testing the plotting.
     """
-
-    filename = DATA_FOLDER_PATH / "problem_outputs.xml"
 
     # First plot
     # This is a rudimentary test as plot are difficult to verify
@@ -52,12 +55,14 @@ def test_wing_geometry_plot():
         fig = wing_geometry_plot(filename, name=f"Plot {i}", fig=fig)
 
 
-def test_aircraft_geometry_plot():
+@pytest.mark.parametrize(
+    "filename",
+    [DATA_FOLDER_PATH / "problem_outputs.xml", DATA_FOLDER_PATH / "problem_outputs_updated.xml"],
+)
+def test_aircraft_geometry_plot(filename):
     """
     Basic tests for testing the plotting.
     """
-
-    filename = DATA_FOLDER_PATH / "problem_outputs.xml"
 
     # First plot
     # This is a rudimentary test as plot are difficult to verify
@@ -76,12 +81,14 @@ def test_aircraft_geometry_plot():
         fig = aircraft_geometry_plot(filename, name=f"Plot {i}", fig=fig)
 
 
-def test_mass_breakdown_bar_plot():
+@pytest.mark.parametrize(
+    "filename",
+    [DATA_FOLDER_PATH / "problem_outputs.xml", DATA_FOLDER_PATH / "problem_outputs_updated.xml"],
+)
+def test_mass_breakdown_bar_plot(filename):
     """
     Basic tests for testing the plotting.
     """
-
-    filename = DATA_FOLDER_PATH / "problem_outputs.xml"
 
     # First plot
     # This is a rudimentary test as plot are difficult to verify
@@ -100,12 +107,14 @@ def test_mass_breakdown_bar_plot():
         _ = mass_breakdown_bar_plot(filename, name=f"Plot {i}", fig=fig)
 
 
-def test_drag_polar_plot():
+@pytest.mark.parametrize(
+    "filename",
+    [DATA_FOLDER_PATH / "problem_outputs.xml", DATA_FOLDER_PATH / "problem_outputs_updated.xml"],
+)
+def test_drag_polar_plot(filename):
     """
     Basic tests for testing the plotting.
     """
-
-    filename = DATA_FOLDER_PATH / "problem_outputs.xml"
 
     # First plot
     # This is a rudimentary test as plot are difficult to verify
@@ -113,12 +122,14 @@ def test_drag_polar_plot():
     _ = drag_polar_plot(filename)
 
 
-def test_mass_breakdown_sun_plot():
+@pytest.mark.parametrize(
+    "filename",
+    [DATA_FOLDER_PATH / "problem_outputs.xml", DATA_FOLDER_PATH / "problem_outputs_updated.xml"],
+)
+def test_mass_breakdown_sun_plot(filename):
     """
     Basic tests for testing the plotting.
     """
-
-    filename = DATA_FOLDER_PATH / "problem_outputs.xml"
 
     # First plot
     # This is a rudimentary test as plot are difficult to verify
@@ -126,12 +137,44 @@ def test_mass_breakdown_sun_plot():
     _ = mass_breakdown_sun_plot(filename)
 
 
-def test_payload_range_plot():
+def test_mass_breakdown_sun_plot_specific_mission():
+    """
+    Test for testing the sun mass breakdown plotting for a specific mission.
+    """
+
+    filename = DATA_FOLDER_PATH / "problem_outputs_multi_mission.xml"
+
+    mission_1 = "evaluation_mission"
+
+    # Plot 1
+    # Specific mission plot
+    f = mass_breakdown_sun_plot(filename, mission_name=mission_1)
+    # f.show()
+
+    mission_2 = "MTOW_mission"
+
+    # Plot 2
+    # Specific mission plot (sizing mission)
+    f = mass_breakdown_sun_plot(filename, mission_name=mission_2)
+    # f.show()
+
+    mission_3 = "not_a_mission_name"
+
+    # Plot 3
+    # Specific mission plot error
+    with pytest.raises(ValueError) as exc_info:
+        _ = mass_breakdown_sun_plot(filename, mission_name=mission_3)
+        assert f"{mission_3}" in str(exc_info.value)
+
+
+@pytest.mark.parametrize(
+    "filename",
+    [DATA_FOLDER_PATH / "problem_outputs.xml", DATA_FOLDER_PATH / "problem_outputs_updated.xml"],
+)
+def test_payload_range_plot(filename):
     """
     Basic tests for testing the plotting.
     """
-
-    filename = DATA_FOLDER_PATH / "problem_outputs.xml"
 
     # First plot
     # This is a rudimentary test as plot are difficult to verify