+++ src/fastoad/gui/tests/test_mission_viewer.py
@@ -57,3 +57,45 @@ def test_mission_viewer():
     with pytest.raises(TypeError):
         filename = DATA_FOLDER_PATH / "valid_sellar.yml"
         mission_viewer.add_mission(filename, name="Mission 2")
+
+
+def test_mission_display():
+    """
+    Basic tests for testing the display of the mission viewer which ensures that altitude and
+    ground distance are displayed by default
+    """
+
+    # Testing with a .csv file
+    filename = DATA_FOLDER_PATH / "flight_points.csv"
+
+    mission_viewer = MissionViewer()
+
+    mission_viewer.add_mission(filename)
+
+    mission_viewer.display()
+
+    assert mission_viewer._x_widget.value == "ground_distance [m]"
+    assert mission_viewer._y_widget.value == "altitude [m]"
+
+    # Testing with a disorganised .csv file since it seeks the right indices
+    filename = DATA_FOLDER_PATH / "flight_points_disorganised.csv"
+
+    mission_viewer = MissionViewer()
+    mission_viewer.add_mission(filename)
+
+    mission_viewer.display()
+
+    assert mission_viewer._x_widget.value == "ground_distance [m]"
+    assert mission_viewer._y_widget.value == "altitude [m]"
+
+    # Testing with a disorganised .csv file and a missing column
+    filename = DATA_FOLDER_PATH / "flight_points_disorganised_missing_column.csv"
+
+    mission_viewer = MissionViewer()
+    mission_viewer.add_mission(filename)
+
+    mission_viewer.display()
+
+    assert mission_viewer._x_widget.value == "consumed_fuel [kg]"
+    # It's missing the ground distance column so it will use the column at index 1
+    assert mission_viewer._y_widget.value == "altitude [m]"