+++ src/fastoad/openmdao/tests/openmdao_sellar_example/disc2.py
@@ -1,4 +1,5 @@
 """Sellar discipline 2"""
+
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
 #  Copyright (C) 2024 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
@@ -12,6 +13,8 @@
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+import numpy as np
+
 from fastoad._utils.sellar.disc2 import BasicDisc2
 
 
@@ -21,6 +24,20 @@ class Disc2(BasicDisc2):
     def setup(self):
         self.add_input(
             "z", val=[5, 2], desc="variable z", units="m**2"
+        )  # for testing non-None units
+        self.add_input("y1", val=1.0, desc="")
+
+        # shape_by_conn=True here to be sure it does not create problems
+        # in input reading.
+        self.add_output("y2", shape_by_conn=True, val=1.0, desc="")
+
+
+class Disc2Bis(BasicDisc2):
+    """An OpenMDAO component to encapsulate Disc2 discipline"""
+
+    def setup(self):
+        self.add_input(
+            "z", val=[np.nan, np.nan], desc="variable z", units="m**2"
         )  # for testing non-None units
         self.add_input("y1", val=1.0, desc="")