+++ CHANGELOG.rst
@@ -2,6 +2,37 @@
 Changelog
 =========
 
+
+Version 1.7.4
+=============
+- Fixed:
+    - Fixed compatibility of ValidityDomainChecker class with OpenMDAO 3.34. (#553)
+
+Version 1.7.3
+=============
+- Added:
+    - Compatibility with Python 3.11. (#538)
+    - Aircraft plot: minor change in naming of geometry variable for horizontal tail (old name still accepted). (#546)
+
+- Fixed:
+    - Fixed validity checker for array variables. (#537)
+    - In mass breakdown bar plot, legend was controlling visibility only for the right-handed plot. (#545 by @aeomath)
+
+Version 1.7.2
+=============
+- Added:
+    - A `fastoad.testing.run_system()` function is now available in public API for component unit test. (#533)
+
+- Modified:
+    - `pathlib.Path` objects are now accepted whenever a file or folder path is expected. (#521, #522, #525)
+    - Enhanced and documented the `CycleGroup` class. (#528)
+
+- Fixed:
+    - Climb was not stopping when start was already over the asked optimal altitude/flight level. (#526)
+    - Fixed links to OpenMDAO doc. (#527)
+    - Fixed behavior when input variables could be added using `model_options`. (#530)
+    - Fixed the variables displayed by default in MissionViewer. (#535)
+
 Version 1.7.1
 =============
 - Added:
@@ -9,7 +40,6 @@ Version 1.7.1
 
 - Fixed:
     - Missions can now be defined without route. (#515)
-
 
 Version 1.7.0
 =============